configuration of a navigation service
