+++ index.js
@@ -1,16 +1,1 @@
-import { Navigation } from "react-native-navigation";
-import {registerScreens} from './src/screens';
-
-registerScreens();
-
-Navigation.events().registerAppLaunchedListener(() => {
-  Navigation.setRoot({
-    root: {
-      component: {
-        name: "App"
-      }
-    }
-  });
-});
-
-
+import './src/App'
