+++ src/App.js
@@ -1,47 +1,14 @@
-import React from 'react'
-import {
-  View,
-  Text,
-  StyleSheet,
-  AsyncStorage
-} from 'react-native'
+import { Navigation } from "react-native-navigation";
+import { registerScreens } from './screens';
 
-import { goToAuth, goHome } from './navigation'
+registerScreens();
 
-
-export default class App extends React.Component {
-  async componentDidMount() {
-    try {
-      const user = 1
-      // const user = await AsyncStorage.getItem('USER_KEY')
-      console.log('user: ', user)
-      if (user) {
-        goHome()
-      } else {
-        goToAuth()
+Navigation.events().registerAppLaunchedListener(() => {
+  Navigation.setRoot({
+    root: {
+      component: {
+        name: "Initializing"
       }
-    } catch (err) {
-      console.log('error: ', err)
-      goToAuth()
     }
-  }
-
-  render() {
-    return (
-      <View style={styles.container}>
-        <Text style={styles.welcome}>Loading...</Text>
-      </View>
-    )
-  }
-}
-
-const styles = StyleSheet.create({
-  welcome: {
-    fontSize: 28
-  },
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center'
-  }
-})
+  });
+});
