+++ src/Home.js
@@ -22,7 +22,7 @@ export default class Home extends React.Component {
   render() {
     return (
       <View style={styles.container}>
-        <Text>Hello from Home screen.</Text>
+        <Text>Hello from Home screen..</Text>
       </View>
     )
   }
